Ignore ESC K VT code emitted by GCC
GCC emits coloured output when it's printing to a tty, and coloured
output is disabled e.g., when the output is piped.
When GCC is called by Dune, it doesn't output ANSI sequences.

However, that behaviour can be forced with the
`-fdiagnostics-color=always` flag. GCC has the particularity that
after each color sequence, it will also emit a ESC K "erase from the
cursor the end of line".

Dune did not support this sequence, and this messes up the output. The
sequence does not seem important; clang doesn't use it, and it
clearing the line means that it would already have been written to,
which seems unlikely. So, this patch simply ignores that sequence.

Closes #3263.

Signed-off-by: Antonin Décimo <antonin@tarides.com>